Added BaseTower and BaseEnemy abstract base classes as well as SimpleEnemy and SimpleTower to work as examples.
BaseTower has the basic functionality of searching for objects in a
range and doing damage to them if they are enemies

BaseEnemy moves intelligently towards the goal and exposes and interface
for other objects to interact with

SimpleTower just does some damage to all emeies

SimpleEnemy dies on damage and destroys itself on death

Further, built up a test scene with a few enemies and a tower to
demonstrate that everything is working. To facilitate this, added a goal
with a DummyGoal script that issues an OnDeath event to enemies
onTriggerEnter.

TODO: Add goalManager and SpawnManger to manage spawning enemies and
managing waves