changed around the was BaseTower handles enemies and added a spawner class
BaseTower now passes an array of BaseEnemy to DoDamage instead of
passing them one at a time, letting the implementer decide what to do
with all the nearby enemies

Overengineered a spawner. Cool design, overly complex. Also not needed
for the final game. Will be removed eventually but will be helpful for
testing in the near future